pyproject.toml
[project]
name = "rapid-ml-ops-demo"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3",
    "metaflow",
    "python-dotenv",
    "torch",
    "torchvision",
    "wandb",

]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
